Move confirmations code to separate file
